Split mixed suites by Cloud support
