+++ packages/test/src/test-integration-workflow-info.ts
@@ -1,4 +1,3 @@
-import { setTimeout as setTimeoutPromise } from 'timers/promises';
 import { randomUUID } from 'crypto';
 import * as workflow from '@temporalio/workflow';
 import type { SearchAttributePair } from '@temporalio/common';
@@ -152,40 +151,6 @@ test("WorkflowInfo().lastFailure contains last run's failure on Workflow Failure
   await worker.runUntil(async () => {
     const lastFailure = await handle.result();
     t.is(lastFailure, 'WorkflowWillFail');
-  });
-});
-
-test('Count workflow executions', async (t) => {
-  const { taskQueue, createWorker, executeWorkflow, startWorkflow } = helpers(t);
-  const worker = await createWorker();
-  const client = t.context.env.client;
-
-  await worker.runUntil(async () => {
-    await Promise.all([
-      // Run 2 workflows that will never complete...
-      startWorkflow(completableWorkflow, { args: [false] }),
-      startWorkflow(completableWorkflow, { args: [false] }),
-
-      // ... and 3 workflows that will complete
-      executeWorkflow(completableWorkflow, { args: [true] }),
-      executeWorkflow(completableWorkflow, { args: [true] }),
-      executeWorkflow(completableWorkflow, { args: [true] }),
-    ]);
-  });
-
-  // FIXME: Find a better way to wait for visibility to stabilize
-  await setTimeoutPromise(1000);
-
-  const actualTotal = await client.workflow.count(`TaskQueue = '${taskQueue}'`);
-  t.deepEqual(actualTotal, { count: 5, groups: [] });
-
-  const actualByExecutionStatus = await client.workflow.count(`TaskQueue = '${taskQueue}' GROUP BY ExecutionStatus`);
-  t.deepEqual(actualByExecutionStatus, {
-    count: 5,
-    groups: [
-      { count: 2, groupValues: [['Running']] },
-      { count: 3, groupValues: [['Completed']] },
-    ],
   });
 });
