+++ packages/test/src/test-local-activities.ts
@@ -1,7 +1,7 @@
 import { firstValueFrom, Subject } from 'rxjs';
 import { Context as ActivityContext } from '@temporalio/activity';
 import { ApplicationFailure, defaultPayloadConverter, WorkflowFailedError } from '@temporalio/client';
-import type { LocalActivityOptions, RetryPolicy } from '@temporalio/common';
+import type { LocalActivityOptions } from '@temporalio/common';
 import { msToNumber } from '@temporalio/common/lib/time';
 import { temporal } from '@temporalio/proto';
 import * as workflow from '@temporalio/workflow';
@@ -533,39 +533,6 @@ export const interceptors: workflow.WorkflowInterceptorsFactory = () => {
     ],
   };
 };
-
-export async function getRetryPolicyFromActivityInfo(
-  retryPolicy: RetryPolicy,
-  fromInsideLocal: boolean
-): Promise<object | undefined> {
-  return await (fromInsideLocal
-    ? workflow.proxyLocalActivities({ startToCloseTimeout: '1m', retry: retryPolicy }).retryPolicy()
-    : workflow.proxyActivities({ startToCloseTimeout: '1m', retry: retryPolicy }).retryPolicy());
-}
-
-test.serial('retryPolicy is set correctly', async (t) => {
-  const { executeWorkflow, createWorker } = helpers(t);
-  const worker = await createWorker({
-    activities: {
-      async retryPolicy(): Promise<object | undefined> {
-        return ActivityContext.current().info.retryPolicy;
-      },
-    },
-  });
-
-  const retryPolicy: RetryPolicy = {
-    backoffCoefficient: 1.5,
-    initialInterval: 2.0,
-    maximumAttempts: 3,
-    maximumInterval: 10.0,
-    nonRetryableErrorTypes: ['nonRetryableError'],
-  };
-
-  await worker.runUntil(async () => {
-    t.deepEqual(await executeWorkflow(getRetryPolicyFromActivityInfo, { args: [retryPolicy, true] }), retryPolicy);
-    t.deepEqual(await executeWorkflow(getRetryPolicyFromActivityInfo, { args: [retryPolicy, false] }), retryPolicy);
-  });
-});
 
 export async function runLocalActivityWithNonLocalActivitiesDisabled(): Promise<string> {
   const { echo } = workflow.proxyLocalActivities({ startToCloseTimeout: '1m' });
